Keep multi-attempt client imports free of node:crypto.
Move sticky assignment hashing to a server-only module and tighten correct-answer letter extraction so phrases like "answer is C" parse correctly.
